Moving easing to vector math
